Refactor Workbench hygiene boundaries
