+++ backend/app/services/import_execution.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import hashlib
 import uuid
 from dataclasses import dataclass
 from typing import Any, Literal
@@ -14,30 +13,21 @@ from app.domain.import_asset_context import (
     canonicalize_occurrence_asset_context as _canonicalize_occurrence_asset_context,
 )
 from app.importers import ImporterParseError, ImporterValidationError, build_importer_registry
-from app.models import (
-    AnalysisRun,
-    AnalysisRunStatus,
-    User,
-)
+from app.models import AnalysisRun, AnalysisRunStatus, User
 from app.repositories import RunRepository
 from app.services.analysis import AnalysisService, WorkbenchAnalysisError
-from app.services.import_artifacts import (
-    resolve_workbench_attack_artifact_path as _resolve_workbench_attack_artifact_path,
-)
-from app.services.import_artifacts import (
-    resolve_workbench_provider_snapshot_path as _resolve_workbench_provider_snapshot_path,
-)
-from app.services.import_artifacts import (
-    validate_attack_import_options as _validate_attack_import_options,
-)
-from app.services.import_errors import ImportServiceError
 from app.services.import_execution_context import (
     _apply_workbench_asset_context,
     _apply_workbench_vex,
-    _parse_errors,
 )
 from app.services.import_execution_failures import (
     raise_analysis_failure as _raise_analysis_failure,
+)
+from app.services.import_execution_parse_failures import (
+    raise_parse_failure as _raise_parse_failure,
+)
+from app.services.import_execution_parse_failures import (
+    raise_sidecar_parse_failure as _raise_sidecar_parse_failure,
 )
 from app.services.import_execution_persistence import _persist_workbench_occurrences
 from app.services.import_execution_summary import (
@@ -45,96 +35,55 @@ from app.services.import_execution_summary import (
     _job_status_entry,
     _record_import_audit,
 )
-from app.services.import_uploads import (
-    has_optional_upload as _has_optional_upload,
+from app.services.import_execution_types import (
+    ImportUploadContent,
+    PreparedImportUpload,
+    ProjectImportUploadRequest,
 )
-from app.services.import_uploads import (
-    ignored_line_count as _ignored_line_count,
+from app.services.import_execution_uploads import (
+    apply_stored_upload_summaries as _apply_stored_upload_summaries,
 )
-from app.services.import_uploads import (
-    normalize_input_type as _normalize_input_type,
+from app.services.import_execution_uploads import (
+    mark_import_run_running as _mark_import_run_running,
 )
-from app.services.import_uploads import (
-    optional_upload_summary as _optional_upload_summary,
+from app.services.import_execution_uploads import (
+    prepare_import_upload as _prepare_import_upload,
 )
-from app.services.import_uploads import (
-    reject_unsafe_upload_filename as _reject_unsafe_upload_filename,
+from app.services.import_execution_uploads import (
+    resolve_import_run as _resolve_import_run,
 )
-from app.services.import_uploads import (
-    sanitize_context_filename as _sanitize_context_filename,
+from app.services.import_execution_uploads import (
+    store_prepared_uploads as _store_prepared_uploads,
 )
-from app.services.import_uploads import (
-    sanitize_filename as _sanitize_filename,
-)
-from app.services.import_uploads import (
-    sanitize_parser_error_message as _sanitize_parser_error_message,
-)
-from app.services.import_uploads import (
-    sanitize_vex_filename as _sanitize_vex_filename,
-)
-from app.services.import_uploads import (
-    store_upload as _store_upload,
-)
-from app.services.import_uploads import (
-    upload_storage_ref as _upload_storage_ref,
-)
-from app.services.import_uploads import (
-    upload_summary as _upload_summary,
-)
-from app.services.import_uploads import (
-    upload_summary_with_path as _upload_summary_with_path,
-)
-from app.services.import_uploads import (
-    validate_aggregate_upload_size as _validate_aggregate_upload_size,
-)
-from app.services.import_uploads import (
-    validate_asset_context_upload as _validate_asset_context_upload,
-)
-from app.services.import_uploads import (
-    validate_input_type as _validate_input_type,
-)
-from app.services.import_uploads import (
-    validate_mime_hint as _validate_mime_hint,
-)
-from app.services.import_uploads import (
-    validate_upload_suffix as _validate_upload_suffix,
-)
-from app.services.import_uploads import (
-    validate_vex_upload as _validate_vex_upload,
-)
+
+__all__ = [
+    "ImportUploadContent",
+    "ProjectImportUploadRequest",
+    "execute_project_import_upload",
+]
 
 
 @dataclass(frozen=True, slots=True)
-class ImportUploadContent:
-    """HTTP-independent upload payload accepted by the import service."""
-
-    filename: str | None
-    content_type: str | None
-    content: bytes
-
-
-@dataclass(frozen=True, slots=True)
-class ProjectImportUploadRequest:
-    """Workbench import request normalized at the route boundary."""
-
+class _ImportFailureContext:
+    session: Session
+    run_repo: RunRepository
+    run: AnalysisRun
+    current_user: User
+    project_id: uuid.UUID
+    job_id: str
+    job_history: list[dict[str, str]]
+    ignored_lines: int
     input_type: str
-    file: ImportUploadContent
-    asset_context_file: ImportUploadContent | None = None
-    vex_file: ImportUploadContent | None = None
-    provider_snapshot_file: str | None = None
-    locked_provider_data: bool = False
-    attack_source: str = "none"
-    attack_mapping_file: str | None = None
-    attack_technique_metadata_file: str | None = None
+    execution_mode: str
 
 
-def _append_job_status(
-    status_history: list[dict[str, str]],
-    status: str,
-) -> list[dict[str, str]]:
-    if status_history and status_history[-1].get("status") == status:
-        return status_history
-    return [*status_history, _job_status_entry(status)]
+def _parse_prepared_upload(prepared: PreparedImportUpload) -> list[Any]:
+    occurrences = build_importer_registry().parse(
+        prepared.input_type,
+        prepared.upload_bytes,
+        filename=prepared.stored_filename,
+    )
+    return [_canonicalize_occurrence_asset_context(item) for item in occurrences]
 
 
 async def execute_project_import_upload(
@@ -149,463 +98,151 @@ async def execute_project_import_upload(
     execution_mode: Literal["request", "background"] = "request",
 ) -> AnalysisRun:
     """Securely upload, normalize, and persist one Workbench import file."""
-    normalized_input_type = _normalize_input_type(upload.input_type)
-    _validate_input_type(normalized_input_type)
-    file = upload.file
-    original_filename = file.filename or "upload"
-    _reject_unsafe_upload_filename(original_filename)
-    stored_filename = _sanitize_filename(original_filename)
-    _validate_upload_suffix(stored_filename, input_type=normalized_input_type)
-    _validate_mime_hint(file.content_type, input_type=normalized_input_type)
-    asset_context_upload = (
-        upload.asset_context_file
-        if _has_optional_upload(
-            upload.asset_context_file.filename if upload.asset_context_file is not None else None
-        )
-        else None
-    )
-    asset_context_original_filename = (
-        asset_context_upload.filename if asset_context_upload is not None else None
-    )
-    asset_context_stored_filename: str | None = None
-    if asset_context_upload is not None and asset_context_original_filename is not None:
-        _reject_unsafe_upload_filename(asset_context_original_filename)
-        asset_context_stored_filename = _sanitize_context_filename(
-            asset_context_original_filename,
-            reserved_filename=stored_filename,
-        )
-        _validate_asset_context_upload(
-            asset_context_stored_filename,
-            asset_context_upload.content_type,
-        )
-    vex_upload = (
-        upload.vex_file
-        if _has_optional_upload(upload.vex_file.filename if upload.vex_file is not None else None)
-        else None
-    )
-    vex_original_filename = vex_upload.filename if vex_upload is not None else None
-    vex_stored_filename: str | None = None
-    if vex_upload is not None and vex_original_filename is not None:
-        _reject_unsafe_upload_filename(vex_original_filename)
-        vex_stored_filename = _sanitize_vex_filename(
-            vex_original_filename,
-            reserved_filenames={stored_filename, asset_context_stored_filename},
-        )
-        _validate_vex_upload(vex_stored_filename, vex_upload.content_type)
-    provider_snapshot_path = _resolve_workbench_provider_snapshot_path(
-        upload.provider_snapshot_file,
-        settings=settings,
-    )
-    attack_mapping_path = _resolve_workbench_attack_artifact_path(
-        upload.attack_mapping_file,
-        settings=settings,
-    )
-    attack_metadata_path = _resolve_workbench_attack_artifact_path(
-        upload.attack_technique_metadata_file,
-        settings=settings,
-    )
-    normalized_attack_source = _validate_attack_import_options(
-        attack_source=upload.attack_source,
-        attack_mapping_path=attack_mapping_path,
-        attack_metadata_path=attack_metadata_path,
-    )
-
-    _validate_aggregate_upload_size(
-        settings=settings,
-        payloads=[
-            file.content,
-            asset_context_upload.content if asset_context_upload is not None else None,
-            vex_upload.content if vex_upload is not None else None,
-        ],
-    )
-    upload_bytes = file.content
-    asset_context_bytes = asset_context_upload.content if asset_context_upload is not None else None
-    vex_bytes = vex_upload.content if vex_upload is not None else None
-    upload_sha256 = hashlib.sha256(upload_bytes).hexdigest()
-    asset_context_sha256 = (
-        hashlib.sha256(asset_context_bytes).hexdigest() if asset_context_bytes is not None else None
-    )
-    vex_sha256 = hashlib.sha256(vex_bytes).hexdigest() if vex_bytes is not None else None
-    ignored_lines = _ignored_line_count(normalized_input_type, upload_bytes)
+    prepared = _prepare_import_upload(upload, settings=settings)
     run_repo = RunRepository(session)
-    job_id = str(uuid.uuid4())
-    job_history = [_job_status_entry("pending")]
-    if existing_run_id is not None:
-        run = run_repo.get_analysis_run(existing_run_id)
-        if run is None:
-            raise ImportServiceError(
-                status_code=404,
-                detail="Analysis run not found",
-            )
-        if run.status not in {AnalysisRunStatus.PENDING, AnalysisRunStatus.RUNNING}:
-            return run
-        existing_job = run.summary_json.get("import_job")
-        if isinstance(existing_job, dict):
-            job_id = str(existing_job.get("id") or job_id)
-            raw_history = existing_job.get("status_history")
-            if isinstance(raw_history, list) and raw_history:
-                job_history = [item for item in raw_history if isinstance(item, dict)]
-    else:
-        run = run_repo.create_analysis_run(
-            project_id=project_id,
-            input_type=normalized_input_type,
-            filename=stored_filename,
-            status=AnalysisRunStatus.PENDING,
-            summary_json={
-                "import_job": _job_payload(
-                    job_id=job_id,
-                    status="pending",
-                    status_history=job_history,
-                    execution_mode=execution_mode,
-                ),
-                "input_upload": _upload_summary(
-                    input_type=normalized_input_type,
-                    original_filename=original_filename,
-                    stored_filename=stored_filename,
-                    content_type=file.content_type,
-                    size_bytes=len(upload_bytes),
-                    sha256=upload_sha256,
-                    path=None,
-                ),
-                "asset_context_upload": _optional_upload_summary(
-                    input_type="asset-context-csv",
-                    original_filename=asset_context_original_filename,
-                    stored_filename=asset_context_stored_filename,
-                    content_type=asset_context_upload.content_type
-                    if asset_context_upload is not None
-                    else None,
-                    size_bytes=len(asset_context_bytes)
-                    if asset_context_bytes is not None
-                    else None,
-                    sha256=asset_context_sha256,
-                    path=None,
-                ),
-                "vex_upload": _optional_upload_summary(
-                    input_type="vex-json",
-                    original_filename=vex_original_filename,
-                    stored_filename=vex_stored_filename,
-                    content_type=vex_upload.content_type if vex_upload is not None else None,
-                    size_bytes=len(vex_bytes) if vex_bytes is not None else None,
-                    sha256=vex_sha256,
-                    path=None,
-                ),
-                "created_findings": 0,
-                "updated_findings": 0,
-                "ignored_lines": ignored_lines,
-                "parse_errors": [],
-            },
-        )
-    upload_path = _store_upload(
+    resolved_run = _resolve_import_run(
+        run_repo=run_repo,
+        project_id=project_id,
+        prepared=prepared,
+        existing_run_id=existing_run_id,
+        execution_mode=execution_mode,
+    )
+    run = resolved_run.run
+    if resolved_run.already_finished:
+        return run
+
+    artifacts = _store_prepared_uploads(
         settings,
         project_id=project_id,
         run_id=run.id,
-        filename=stored_filename,
-        content=upload_bytes,
+        prepared=prepared,
     )
-    asset_context_path = (
-        _store_upload(
-            settings,
-            project_id=project_id,
-            run_id=run.id,
-            filename=asset_context_stored_filename or "asset_context.csv",
-            content=asset_context_bytes,
-        )
-        if asset_context_bytes is not None
-        else None
+    _apply_stored_upload_summaries(
+        run,
+        resolved_run=resolved_run,
+        artifacts=artifacts,
+        execution_mode=execution_mode,
     )
-    vex_path = (
-        _store_upload(
-            settings,
-            project_id=project_id,
-            run_id=run.id,
-            filename=vex_stored_filename or "openvex.json",
-            content=vex_bytes,
-        )
-        if vex_bytes is not None
-        else None
-    )
-    upload_ref = _upload_storage_ref(
-        project_id=project_id,
-        run_id=run.id,
-        filename=stored_filename,
-    )
-    asset_context_ref = (
-        _upload_storage_ref(
-            project_id=project_id,
-            run_id=run.id,
-            filename=asset_context_stored_filename or "asset_context.csv",
-        )
-        if asset_context_path is not None
-        else None
-    )
-    vex_ref = (
-        _upload_storage_ref(
-            project_id=project_id,
-            run_id=run.id,
-            filename=vex_stored_filename or "openvex.json",
-        )
-        if vex_path is not None
-        else None
-    )
-    run.summary_json = {
-        **run.summary_json,
-        "import_job": _job_payload(
-            job_id=job_id,
-            status="pending",
-            status_history=job_history,
-            execution_mode=execution_mode,
-        ),
-        "input_upload": {
-            **run.summary_json["input_upload"],
-            "path": upload_ref,
-            "storage_ref": upload_ref,
-        },
-        "asset_context_upload": _upload_summary_with_path(
-            run.summary_json.get("asset_context_upload"),
-            path=asset_context_ref,
-        ),
-        "vex_upload": _upload_summary_with_path(
-            run.summary_json.get("vex_upload"),
-            path=vex_ref,
-        ),
-    }
     if defer_execution:
         session.commit()
         session.refresh(run)
         return run
 
-    run.status = AnalysisRunStatus.RUNNING
-    job_history = _append_job_status(job_history, "running")
-    run.summary_json = {
-        **run.summary_json,
-        "import_job": _job_payload(
-            job_id=job_id,
-            status="running",
-            status_history=job_history,
-            execution_mode=execution_mode,
-        ),
-    }
+    job_history = _mark_import_run_running(
+        run,
+        job_id=resolved_run.job_id,
+        job_history=resolved_run.job_history,
+        execution_mode=execution_mode,
+    )
     session.flush()
+    failure_context = _ImportFailureContext(
+        session=session,
+        run_repo=run_repo,
+        run=run,
+        current_user=current_user,
+        project_id=project_id,
+        job_id=resolved_run.job_id,
+        job_history=job_history,
+        ignored_lines=prepared.ignored_lines,
+        input_type=prepared.input_type,
+        execution_mode=execution_mode,
+    )
 
     try:
-        occurrences = build_importer_registry().parse(
-            normalized_input_type,
-            upload_bytes,
-            filename=stored_filename,
-        )
-        occurrences = [_canonicalize_occurrence_asset_context(item) for item in occurrences]
+        occurrences = _parse_prepared_upload(prepared)
     except (ImporterParseError, ImporterValidationError) as exc:
-        parse_errors = _parse_errors(
-            exc, filename=stored_filename, input_type=normalized_input_type
+        _raise_parse_failure(
+            session=failure_context.session,
+            run_repo=failure_context.run_repo,
+            run=failure_context.run,
+            current_user=failure_context.current_user,
+            project_id=failure_context.project_id,
+            job_id=failure_context.job_id,
+            job_history=failure_context.job_history,
+            ignored_lines=failure_context.ignored_lines,
+            input_type=failure_context.input_type,
+            filename=prepared.stored_filename,
+            exc=exc,
+            execution_mode=failure_context.execution_mode,
         )
-        failed_history = [*job_history, _job_status_entry("failed")]
-        failed_run = run_repo.finish_analysis_run(
-            run.id,
-            status=AnalysisRunStatus.FAILED,
-            error_message=_sanitize_parser_error_message(str(exc)),
-            error_json={
-                "parse_errors": parse_errors,
-                "created_findings": 0,
-                "updated_findings": 0,
-                "ignored_lines": ignored_lines,
-                "import_job": _job_payload(
-                    job_id=job_id,
-                    status="failed",
-                    status_history=failed_history,
-                    execution_mode=execution_mode,
-                ),
-            },
-            summary_json={
-                **run.summary_json,
-                "import_job": _job_payload(
-                    job_id=job_id,
-                    status="failed",
-                    status_history=failed_history,
-                    execution_mode=execution_mode,
-                ),
-                "parse_errors": parse_errors,
-                "created_findings": 0,
-                "updated_findings": 0,
-                "ignored_lines": ignored_lines,
-            },
-        )
-        _record_import_audit(
-            session,
-            current_user=current_user,
-            project_id=project_id,
-            run_id=failed_run.id,
-            status="failure",
-            stage="parse",
-            input_type=normalized_input_type,
-        )
-        session.commit()
-        raise ImportServiceError(
-            status_code=422,
-            detail={
-                "message": "Import parsing failed.",
-                "analysis_run_id": str(failed_run.id),
-                "ignored_lines": ignored_lines,
-                "parse_errors": parse_errors,
-            },
-        ) from exc
 
     asset_context_summary: dict[str, Any] | None = None
-    if asset_context_path is not None:
+    if artifacts.asset_context_path is not None:
         try:
             occurrences, asset_context_summary = _apply_workbench_asset_context(
                 occurrences,
-                asset_context_path=asset_context_path,
+                asset_context_path=artifacts.asset_context_path,
             )
         except ValueError as exc:
-            error_message = _sanitize_parser_error_message(str(exc))
-            asset_context_error = {
-                "message": error_message,
-                "filename": asset_context_stored_filename,
-                "stage": "asset_context_parse",
-                "error_type": exc.__class__.__name__,
-            }
-            failed_history = [*job_history, _job_status_entry("failed")]
-            failed_run = run_repo.finish_analysis_run(
-                run.id,
-                status=AnalysisRunStatus.FAILED,
-                error_message=error_message,
-                error_json={
-                    "asset_context_error": asset_context_error,
-                    "created_findings": 0,
-                    "updated_findings": 0,
-                    "ignored_lines": ignored_lines,
-                    "import_job": _job_payload(
-                        job_id=job_id,
-                        status="failed",
-                        status_history=failed_history,
-                        execution_mode=execution_mode,
-                    ),
-                },
-                summary_json={
-                    **run.summary_json,
-                    "import_job": _job_payload(
-                        job_id=job_id,
-                        status="failed",
-                        status_history=failed_history,
-                        execution_mode=execution_mode,
-                    ),
-                    "asset_context_error": asset_context_error,
-                    "parse_errors": [],
-                    "created_findings": 0,
-                    "updated_findings": 0,
-                    "ignored_lines": ignored_lines,
-                },
-            )
-            _record_import_audit(
-                session,
-                current_user=current_user,
-                project_id=project_id,
-                run_id=failed_run.id,
-                status="failure",
+            _raise_sidecar_parse_failure(
+                session=failure_context.session,
+                run_repo=failure_context.run_repo,
+                run=failure_context.run,
+                current_user=failure_context.current_user,
+                project_id=failure_context.project_id,
+                job_id=failure_context.job_id,
+                job_history=failure_context.job_history,
+                ignored_lines=failure_context.ignored_lines,
+                input_type=failure_context.input_type,
+                error_key="asset_context_error",
+                response_message="Asset context parsing failed.",
+                filename=prepared.asset_context.stored_filename,
                 stage="asset_context_parse",
-                input_type=normalized_input_type,
+                exc=exc,
+                execution_mode=failure_context.execution_mode,
             )
-            session.commit()
-            raise ImportServiceError(
-                status_code=422,
-                detail={
-                    "message": "Asset context parsing failed.",
-                    "analysis_run_id": str(failed_run.id),
-                    "asset_context_error": asset_context_error,
-                },
-            ) from exc
 
     vex_summary: dict[str, Any] | None = None
-    if vex_path is not None:
+    if artifacts.vex_path is not None:
         try:
             occurrences, vex_summary = _apply_workbench_vex(
                 occurrences,
-                vex_path=vex_path,
+                vex_path=artifacts.vex_path,
             )
         except ValueError as exc:
-            error_message = _sanitize_parser_error_message(str(exc))
-            vex_error = {
-                "message": error_message,
-                "filename": vex_stored_filename,
-                "stage": "vex_parse",
-                "error_type": exc.__class__.__name__,
-            }
-            failed_history = [*job_history, _job_status_entry("failed")]
-            failed_run = run_repo.finish_analysis_run(
-                run.id,
-                status=AnalysisRunStatus.FAILED,
-                error_message=error_message,
-                error_json={
-                    "vex_error": vex_error,
-                    "created_findings": 0,
-                    "updated_findings": 0,
-                    "ignored_lines": ignored_lines,
-                    "import_job": _job_payload(
-                        job_id=job_id,
-                        status="failed",
-                        status_history=failed_history,
-                        execution_mode=execution_mode,
-                    ),
-                },
-                summary_json={
-                    **run.summary_json,
-                    "import_job": _job_payload(
-                        job_id=job_id,
-                        status="failed",
-                        status_history=failed_history,
-                        execution_mode=execution_mode,
-                    ),
-                    "vex_error": vex_error,
-                    "parse_errors": [],
-                    "created_findings": 0,
-                    "updated_findings": 0,
-                    "ignored_lines": ignored_lines,
-                },
-            )
-            _record_import_audit(
-                session,
-                current_user=current_user,
-                project_id=project_id,
-                run_id=failed_run.id,
-                status="failure",
+            _raise_sidecar_parse_failure(
+                session=failure_context.session,
+                run_repo=failure_context.run_repo,
+                run=failure_context.run,
+                current_user=failure_context.current_user,
+                project_id=failure_context.project_id,
+                job_id=failure_context.job_id,
+                job_history=failure_context.job_history,
+                ignored_lines=failure_context.ignored_lines,
+                input_type=failure_context.input_type,
+                error_key="vex_error",
+                response_message="VEX parsing failed.",
+                filename=prepared.vex.stored_filename,
                 stage="vex_parse",
-                input_type=normalized_input_type,
+                exc=exc,
+                execution_mode=failure_context.execution_mode,
             )
-            session.commit()
-            raise ImportServiceError(
-                status_code=422,
-                detail={
-                    "message": "VEX parsing failed.",
-                    "analysis_run_id": str(failed_run.id),
-                    "vex_error": vex_error,
-                },
-            ) from exc
 
     try:
         analysis_result = AnalysisService(session, settings).analyze_import(
-            input_path=upload_path,
-            input_type=normalized_input_type,
-            asset_context_file=asset_context_path,
-            provider_snapshot_file=provider_snapshot_path,
-            locked_provider_data=upload.locked_provider_data,
-            attack_source=normalized_attack_source,
-            attack_mapping_file=attack_mapping_path,
-            attack_technique_metadata_file=attack_metadata_path,
-            vex_files=[vex_path] if vex_path is not None else [],
+            input_path=artifacts.upload_path,
+            input_type=prepared.input_type,
+            asset_context_file=artifacts.asset_context_path,
+            provider_snapshot_file=prepared.provider_snapshot_path,
+            locked_provider_data=prepared.locked_provider_data,
+            attack_source=prepared.attack_source,
+            attack_mapping_file=prepared.attack_mapping_path,
+            attack_technique_metadata_file=prepared.attack_metadata_path,
+            vex_files=[artifacts.vex_path] if artifacts.vex_path is not None else [],
         )
     except WorkbenchAnalysisError as exc:
         _raise_analysis_failure(
-            session=session,
-            run_repo=run_repo,
-            run=run,
-            current_user=current_user,
-            project_id=project_id,
-            job_id=job_id,
-            job_history=job_history,
-            ignored_lines=ignored_lines,
-            input_type=normalized_input_type,
+            session=failure_context.session,
+            run_repo=failure_context.run_repo,
+            run=failure_context.run,
+            current_user=failure_context.current_user,
+            project_id=failure_context.project_id,
+            job_id=failure_context.job_id,
+            job_history=failure_context.job_history,
+            ignored_lines=failure_context.ignored_lines,
+            input_type=failure_context.input_type,
             exc=exc,
-            execution_mode=execution_mode,
+            execution_mode=failure_context.execution_mode,
         )
 
     persist_summary = _persist_workbench_occurrences(
@@ -622,7 +259,7 @@ async def execute_project_import_upload(
         summary_json={
             **run.summary_json,
             "import_job": _job_payload(
-                job_id=job_id,
+                job_id=resolved_run.job_id,
                 status="succeeded",
                 status_history=[*job_history, _job_status_entry("succeeded")],
                 execution_mode=execution_mode,
@@ -631,8 +268,8 @@ async def execute_project_import_upload(
             **persist_summary,
             "asset_context": asset_context_summary,
             "vex": vex_summary,
-            "ignored_lines": ignored_lines,
-            "input_sha256": upload_sha256,
+            "ignored_lines": prepared.ignored_lines,
+            "input_sha256": prepared.upload_sha256,
             "parse_errors": [],
         },
     )
@@ -643,7 +280,7 @@ async def execute_project_import_upload(
         run_id=finished_run.id,
         status="success",
         stage="succeeded",
-        input_type=normalized_input_type,
+        input_type=prepared.input_type,
     )
     session.commit()
     return finished_run
