+++ frontend/src/components/assets/asset-format-model.ts
@@ -1,0 +1,13 @@
+export function formatDateTime(value: string | null | undefined) {
+  if (!value) {
+    return "N.A."
+  }
+  const date = new Date(value)
+  if (Number.isNaN(date.getTime())) {
+    return "N.A."
+  }
+  return new Intl.DateTimeFormat(undefined, {
+    dateStyle: "medium",
+    timeStyle: "short",
+  }).format(date)
+}
